Search the Bible

This will return results based on your specific search query

bible_search(
  query = NULL,
  limit = 10,
  offset = NULL,
  sort = "relevance",
  range = NULL,
  fuzziness = "AUTO",
  bibleid = Sys.getenv("MAIN_BIBLEID"),
  debug = FALSE,
  apikey = Sys.getenv("BIBLER_APIKEY")
)

Arguments

query
Search keywords or passage reference. Supported wildcards are * and ?.The * wildcard matches any character sequence (e.g. searching for "wo*d" finds text such as "word", "world", and "worshipped"). The ? wildcard matches any matches any single character (e.g. searching for "l?ve" finds text such as "live" and "love").

limit
Integer limit for how many matching results to return. Default is 10.

offset
Offset for search results. Used to paginate results

sort
Sort order of results. Supported values are relevance (default), canonical and reverse-canonical

range
One or more, comma separated, passage ids (book, chapter, verse) which the search will be limited to. (i.e. gen.1,gen.5 or gen-num or gen.1.1-gen.3.5)

fuzziness
Sets the fuzziness of a search to account for misspellings. Values can be 0, 1, 2, or AUTO. Defaults to AUTO which varies depending on the

bibleid
Identifies the Bible version (required)

debug
Used to help debug the query

apikey
API.bible api key. Sign up here https://scripture.api.bible/signup